refactoring: relocate snapshot tests to detached folder
