+++ src/tests/snapshot/task-edit.test.js
@@ -1,5 +1,5 @@
-import TaskEditView from '../view/task-edit';
-import {Color, RepeatingDaysMap} from '../const';
+import TaskEditView from '../../view/task-edit';
+import {Color, RepeatingDaysMap} from '../../const';
 
 it(`TaskEditor rendering`, () => {
   const task = {
